+++ game-server/src/main/java/com/kingston/mmorpg/game/Modules.java
@@ -3,17 +3,17 @@ package com.kingston.mmorpg.game;
 public interface Modules {
 
 	// ------------------底层功能支持模块（-128 ~ -1）-----------------
-	int GM = -1;
+	short GM = -1;
 
-	// ------------------业务功能模块（1~127）---------------------
+	// ------------------业务功能模块（1~326）---------------------
 
 	/** 玩家 */
-	byte PLAYER = 1;
+	short PLAYER = 1;
 	/** 场景 */
-	byte SCENE = 2;
+	short SCENE = 2;
 	/** 活动 */
-	byte ACTIVITY = 3;
+	short ACTIVITY = 3;
 	/** 技能 */
-	byte SKILL = 4;
+	short SKILL = 4;
 
 }
